requirements.txt
fastapi
uvicorn
python-multipart
pillow
pillow-heif
anthropic
python-dotenv